+++ .editorconfig
@@ -21,3 +21,6 @@ indent_size = 2
 indent_style = space
 indent_size = 2
 
+[*.scss,*.css]
+indent_style = space
+indent_size = 2
